rpmsg: smd: Add quirk to initiate channel open from AP
RPM Firmware on variety of newer SoCs (MSM8917, MSM8937, MSM8953,
SDM630, SDM660 and on some MSM8916 devices) doesn't initiate opening of
the SMD channel.  This doesn't allow probing of smd-rpm driver because
there is a check that requires RPM side of channel to be opened first.

This commit adds a property that signals SMD driver that it should open
AP side of channel first for specific channel. With this flag if RPM
doesn't open its side of channel in responce driver won't continue with
creation of client device.

Another option is to remove the check, but it's more controversial.
Changing signed firmware on thousands of devices isn't really an option.

Signed-off-by: Vladimir Lypak <junak.pub@gmail.com>