net: ipa: Add timeout for ipa_cmd_pipeline_clear_wait
Sometimes the pipeline clear fails, and when it does, having a hang in
kernel is ugly. The timeout gives us a nice error message. Note that
this shouldn't actually hang, ever. It only hangs if there is a mistake
in the config, and the timeout is only useful when debugging.

Signed-off-by: Vladimir Lypak <vladimir.lypak@gmail.com>